refactor(tibuild): rename `internalpkg` to `internal`
- rename the pkg name.
- extract complex steps of `PipelineTrigger` to `triggerJenkinsJob`.
- simplfy the gin bind steps.
